chore: clean unused parts, move some functions to utils
