+++ packages/utils/index.ts
@@ -4,3 +4,4 @@ export * from './address'
 export * from './grpc'
 export * from './types'
 export * from './tx'
+export * from './solana'
